refact: extract method
